+++ example/lib/mixins/debounced_slider_mixin.dart
@@ -1,0 +1,40 @@
+import 'dart:async';
+
+import 'package:flutter/widgets.dart';
+
+mixin DebouncedSliderMixin<T extends StatefulWidget> on State<T> {
+  double _value = 20;
+  double _renderedValue = 20;
+  Timer? _debounce;
+
+  @protected
+  double get initialSliderValue => 20;
+
+  @protected
+  Duration get sliderDebounce => const Duration(milliseconds: 150);
+
+  double get sliderValue => _value;
+
+  double get renderedValue => _renderedValue;
+
+  @override
+  void initState() {
+    super.initState();
+    _value = initialSliderValue;
+    _renderedValue = initialSliderValue;
+  }
+
+  @override
+  void dispose() {
+    _debounce?.cancel();
+    super.dispose();
+  }
+
+  void onSliderChanged(double v) {
+    setState(() => _value = v);
+    _debounce?.cancel();
+    _debounce = Timer(sliderDebounce, () {
+      if (mounted) setState(() => _renderedValue = v);
+    });
+  }
+}